Add files via upload
A website that has over 1700 reviews of chocolate bars from all around the world. It’s displayed in the web browser on this page.
To figure out: Where are the best cocoa beans grown? Which countries produce the highest-rated bars? What’s the relationship between cocoa solids percentage and rating?